+++ test/test_fitprior.jl
@@ -4,7 +4,6 @@
 @test_nothrow fitprior(pnts, "k-means", false; useR = false, verbose = false) # kmeans with points, Julia
 @test_nothrow fitprior(pnts, "k-medoids", false; useR = false, verbose = false) # kmedoids with points, Julia
 @test_nothrow fitprior(distM, "k-medoids", true; useR = false, verbose = false) # kmedoids with distances, Julia
-params = ()
 @test_logs (:warn,) global params = fitprior(pnts, "k-means", false; useR = false, verbose = false, Kmin = 1, Kmax = 1) # check edge case Kmax = 2
 @test params.K_initial == 1
 @test_logs (:warn,) global params = fitprior(pnts, "k-means", false; useR = false, verbose = false, Kmin = N, Kmax = N) # check edge case Kmin = N-1
@@ -15,3 +14,5 @@ params = ()
 @test_throws ArgumentError fitprior(pnts, "k-means", false; Kmin = 0, verbose = false) # check 1 ≤ Kmin
 @test_throws ArgumentError fitprior(pnts, "k-means", false; Kmin = N, Kmax = 1, verbose = false) # check Kmin ≤ Kmax
 @test_throws ArgumentError fitprior(pnts, "k-means", false; Kmax = length(pnts) + 1, verbose = false) # check Kmax ≤ N
+@test_nothrow show(params)
+@test_nothrow display(params)