+++ test/test_pointestimates.jl
@@ -6,6 +6,7 @@ temp = sample(1:K, N)
 @test binderloss(temp, temp; normalised = true) ≈ 0 atol = 1e-9
 @test_throws ArgumentError binderloss(temp, [temp; 1]) # check that lengths are equal
 @test binderloss(temp, temp; normalised = false) ≈ 0 atol = 1e-9
+# test the methods that should work
 @test_nothrow getpointestimate(result; method = "MAP")
 @test_nothrow getpointestimate(result; method = "MLE")
 @test_nothrow getpointestimate(result; loss = "binder", method = "MPEL")
@@ -16,5 +17,9 @@ temp = sample(1:K, N)
 @test_nothrow getpointestimate(result; loss = "omARI", method = "SALSO")
 @test_nothrow getpointestimate(result; loss = "VI", method = "SALSO")
 @test_nothrow getpointestimate(result; loss = "ID", method = "SALSO")
+@test_nothrow getpointestimate(result; loss = varinfo, method = "MPEL")
+# test bad inputs
+@test_throws ArgumentError getpointestimate(result; loss = (x, y) -> 1.0, method = "SALSO")
 @test_throws ArgumentError getpointestimate(result; loss = "ID", method = "some other method")
 @test_throws ArgumentError getpointestimate(result; loss = "some other loss function", method = "SALSO")
+@test_throws ArgumentError getpointestimate(result; loss = "some other loss function", method = "MPEL")